lactf-challenges-2025/web/i-spy/app/requirements.txt
flask
pyjwt[crypto]
gunicorn